requirements.txt
backports.tarfile==1.2.0
boto3==1.38.3
dj-database-url==2.3.0
django-storages==1.14.6
django-tinymce==4.1.0
django-two-factor-auth==1.17.0
environs==14.1.1
gunicorn==23.0.0
importlib-metadata==8.0.0
jaraco.collections==5.1.0
libsass==0.23.0
lxml==5.3.2
phonenumbers==9.0.5
pillow==11.1.0
platformdirs==4.3.6
psycopg2-binary==2.9.10
tomli==2.0.1
whitenoise==6.9.0